libctr/init_linux: reorder chdir
commit 910e741c245dec82bf72388b15389d0427c56c6f moved the chdir to be one of the
first steps of finalizing the namespace of the container.

However, this causes issues when the cwd is not accessible by the user running runc, but rather
as the container user.

Thus, setupUser has to happen before we call chdir. setupUser still happens before setting the caps,
so the user should be privileged enough to mitigate the issues fixed in 910e741c245dec82bf72388b15389d0427c56c6f

Signed-off-by: Peter Hunt <pehunt@redhat.com>