libct: document process.LogLevel field
This field used to hold a string representation of log level (like
"debug" or "info"). Since commit 5bf60ee942fc2c9 this is now a string
holding a numeric representation of log level (e.g. "4"). This was done
in preparation to commit c3c6784aacbf7f78, and simplifies things in
a few places.

Let's document it.

Fixes: 5bf60ee942fc2c9
Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>