Revert "increase memory.max in cgroups.bats"
This reverts commit 5967347e02941a0a63206a4e9d15517ec1eecc7a.

We needed [1] because when we removed the bindfd logic in [2] we had not
yet moved the binary cloning logic to Go and thus it was necessary to
increase the memory limit in CI because the clone was happening after
joining the cgroup. However, [3] finally moved that code to Go and thus
the cloning is now done outside of the container's cgroup and thus is no
longer accounted as part of the container's memory usage at any point.

Now we can properly support running a simple container with lower memory
usage as we did before.

[1]: commit 5967347e0294 ("increase memory.max in cgroups.bats")
[2]: commit 4d6021363cdd ("nsenter: cloned_binary: remove bindfd logic entirely")
[3]: commit dbbcd5f3879b ("nsexec: migrate memfd /proc/self/exe logic to Go code")

Signed-off-by: lfbzhm <lifubang@acmcoder.com>
[cyphar: fixed commit messages]
Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>