tests/int/cgroups.bats: exclude dmem controller
The dmem controller is added into kernel v6.13 and is now enabled in
Fedora 42 kernels. Yet, systemd is not aware of dmem.

This fixes the test case failure on Fedora.

For the initial test case, see commit 75079c81.

For earlier commits similar to this one, see
commits 96eca30c, 2c955204, a3f717ae.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>